feat(models): prefer a single types file
