+++ models/payment/src/payment-method-info/payment-method-info-input/builder.spec.ts
@@ -4,10 +4,38 @@ import { createBuilderSpec } from '@commercetools-test-data/core/test-utils';
 import {
   TPaymentMethodInfoInput,
   TPaymentMethodInfoInputGraphQL,
-} from './types';
+} from '../types';
 import * as PaymentMethodInfoInput from './index';
 
 describe('builder', () => {
+  it(
+    ...createBuilderSpec<TPaymentMethodInfoInput, TPaymentMethodInfoInput>(
+      'default',
+      PaymentMethodInfoInput.random(),
+      expect.objectContaining({
+        paymentInterface: null,
+        method: null,
+        name: expect.objectContaining({
+          en: expect.any(String),
+        }),
+      })
+    )
+  );
+
+  it(
+    ...createBuilderSpec<TPaymentMethodInfoInput, TPaymentMethodInfoInput>(
+      'rest',
+      PaymentMethodInfoInput.random(),
+      expect.objectContaining({
+        paymentInterface: null,
+        method: null,
+        name: expect.objectContaining({
+          en: expect.any(String),
+        }),
+      })
+    )
+  );
+
   it(
     ...createBuilderSpec<
       TPaymentMethodInfoInput,
