+++ models/payment/src/payment-method-info/payment-method-info-input/builder.ts
@@ -1,10 +1,10 @@
 import { Builder } from '@commercetools-test-data/core';
-import generator from './generator';
-import transformers from './transformers';
 import type {
   TPaymentMethodInfoInput,
   TCreatePaymentMethodInfoInputBuilder,
-} from './types';
+} from '../types';
+import generator from './generator';
+import transformers from './transformers';
 
 const Model: TCreatePaymentMethodInfoInputBuilder = () =>
   Builder<TPaymentMethodInfoInput>({
