+++ models/payment/src/payment-method-info/payment-method-info-input/index.ts
@@ -1,3 +1,3 @@
 export { default as random } from './builder';
 export { default as presets } from './presets';
-export * from './types';
+export * from '../types';
