+++ models/payment/src/payment-method-info/types.ts
@@ -9,5 +9,15 @@ export type TPaymentMethodInfoGraphql = TPaymentMethodInfo & {
   __typename: 'PaymentMethodInfo';
 };
 
+export type TPaymentMethodInfoInput = PaymentMethodInfo;
+
+export type TPaymentMethodInfoInputGraphQL = TPaymentMethodInfoInput & {
+  __typename: 'PaymentMethodInfoInput';
+};
+
 export type TPaymentMethodInfoBuilder = TBuilder<TPaymentMethodInfo>;
 export type TCreatePaymentMethodInfoBuilder = () => TPaymentMethodInfoBuilder;
+
+export type TPaymentMethodInfoInputBuilder = TBuilder<TPaymentMethodInfoInput>;
+export type TCreatePaymentMethodInfoInputBuilder =
+  () => TPaymentMethodInfoInputBuilder;
